requirements.txt
## version doesnt matter
tqdm
opencv-python

tensorflow-gpu ~= 2.3.0
